Sunday instructions and scripts work. (#25)
* Update mm-install-system-updates.sh

* Update -mm-mount-mmojo-share.sh

* Update mm-environment-variables.sh

* Update mm-main-branch.sh

* Update mm-show-build-tools.sh

* Update mm-update-local-mmojo-server-repo.sh

* Update mm-work-in-progress-branch.sh

* Update 300-Gather-Build-Pieces.md

* Update 301-Create-Models-Directory.sh

* Update 304-Clone-Build-openssl.sh

* Update 304-Clone-Build-openssl.sh

* Update 304-Clone-Build-openssl.sh

* Update 404-Assemble-Cosmo-APE-Local.sh

* Update mm-environment-variables.sh

* Update 403-Test-Debug-Command-Line.sh

* Update 403-Test-Debug-mmojo-server-support.sh

* Update 403-Test-Debug-mmojo-server-args.sh

* Update 403-Test-Debug-Command-Line.sh

* Update 402-Prepare-to-Test.md

* Update 403-Test-Debug-Command-Line.sh

* Update 403-Test-Debug-mmojo-server-args.sh

* Update 403-Test-Debug-mmojo-server-support.sh

* Update 404-Test-Cosmo-APE.sh

* Update 404-Test-Cosmo-aarch64.sh

* Update 404-Test-Cosmo-x86_64.sh

* Update 405-Test-CPU.sh

* Update 406-Test-CUDA.sh

* Update 407-Test-Vulkan.sh

* Update 402-Prepare-to-Test.md

* Update mm-environment-variables.sh

* Update mm-environment-variables.sh

* Update 403-Test-Debug-Command-Line.sh

* Update 402-Prepare-to-Test.md

* Update 403-Test-Debug-Command-Line.sh

* Update 403-Test-Debug-mmojo-server-args.sh

* Update 403-Test-Debug-mmojo-server-support.sh

* Update 404-Test-Cosmo-APE.sh

* Update 404-Test-Cosmo-aarch64.sh

* Update 404-Test-Cosmo-x86_64.sh

* Update 403-Test-Debug-Command-Line.sh

* Update 405-Test-CPU.sh

* Update 406-Test-CUDA.sh

* Update 407-Test-Vulkan.sh

* Update 402-Prepare-to-Test.md

+++ scripts/301-Create-Models-Directory.sh
@@ -14,10 +14,12 @@ printf "\n**********\n*\n* STARTED: $SCRIPT_NAME.\n*\n**********\n\n"
 mkdir -p $MODELS_DIR
 cd $MODELS_DIR
 
+if ! test -f "$MODEL_MAP"; then
 cat << EOF > $MODEL_MAP
 Google-Gemma-1B-Instruct-v3-q8_0.gguf mmojo-server-Google-Gemma-1B-Instruct-v3
 Google-Gemma-4B-Instruct-v3-q8_0.gguf mmojo-server-Google-Gemma-4B-Instruct-v3
 EOF
+fi
 
 cd $HOME
